Bear walks, and hits the player
Missing: clean up the "Walk" script. Too many unused comments.
Missing: death animation. Need to use the same plugin as for the skull when dead. Easy to do, but need to do it.
Missing: Stomp: we need to define the spawn space.
Missing: increase the slash collider. It's even missing this new collider.
Missing: throw rocks. It has the spawn point but the script is missing.

So far so good!!!